Move generalize_sql to helpers
