+++ reporter/test/test_php_errors_source.py
@@ -129,29 +129,6 @@ class DBErrorsSourceTestClass(unittest.TestCase):
     def setUp(self):
         self._source = DBQueryErrorsSource()
 
-    def test_generalize_sql(self):
-        assert DBQueryErrorsSource._generalize_sql(None) is None
-
-        assert DBQueryErrorsSource._generalize_sql("UPDATE  `category` SET cat_pages = cat_pages + 1,cat_files = cat_files + 1 WHERE cat_title = 'foo'") ==\
-            "UPDATE `category` SET cat_pages = cat_pages + N,cat_files = cat_files + N WHERE cat_title = X"
-
-        assert DBQueryErrorsSource._generalize_sql("SELECT  entity_key  FROM `wall_notification_queue`  WHERE (wiki_id = ) AND (event_date > '20150105141012')") ==\
-            "SELECT entity_key FROM `wall_notification_queue` WHERE (wiki_id = ) AND (event_date > X)"
-
-        assert DBQueryErrorsSource._generalize_sql("UPDATE  `user` SET user_touched = '20150112143631' WHERE user_id = '25239755'") ==\
-            "UPDATE `user` SET user_touched = X"
-
-        # multiline query
-        sql = """
-SELECT page_title
-    FROM page
-    WHERE page_namespace = '10'
-    AND page_title COLLATE LATIN1_GENERAL_CI LIKE '%{{Cata%'
-        """
-
-        assert DBQueryErrorsSource._generalize_sql(sql) ==\
-            "SELECT page_title FROM page WHERE page_namespace = X AND page_title COLLATE LATINN_GENERAL_CI LIKE X"
-
     def test_get_context_from_entry(self):
         context = DBQueryErrorsSource._get_context_from_entry({
             "@exception": {
